Fixing the misaligned store at compiler_x86.c:112 let the JIT compiler
run past it, and -fno-sanitize-recover=all then aborted on the left
shift of a negative displacement further down the same function. Build 3
reported it four times and nothing else: trunnel.c:207 and
circuitbuild.c:1014 were both gone.

+++ src/ext/equix/hashx/src/compiler_x86.c
@@ -149,7 +149,7 @@ bool hashx_compile_x86(const hashx_program* program, uint8_t* code) {
 			break;
 		case INSTR_BRANCH:
 			EMIT_U64(pos, ((uint64_t)instr->imm32) << 32 | 0xc2f7f209);
-			EMIT_U16(pos, ((target - pos) << 8) | 0x74);
+			EMIT_U16(pos, ((uint32_t)(target - pos) << 8) | 0x74);
 			break;
 		default:
 			UNREACHABLE;
